requirements.txt
requests~=2.27.0
python-dotenv~=0.18.0
sendgrid~=6.7.1
lemon~=1.0.3